moved image related utils to image_utils
